# src/PDMS/Product/Server/PdmsFrontdoor/OpenApiDocument.OpenApi3_0.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PdmsFrontdoor",
    "version": "0.0.851.2"
  },
  "servers": [
    {
      "url": "https://management.privacy.microsoft.com"
    }
  ],
  "paths": {
...
    "/api/v2/TransferRequests('{id}')": {
      "delete": {
        "tags": [
          "TransferRequest V2"
        ],
        "summary": "Deletes a Transfer request by id.",
        "operationId": "deleteApiV2ById",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authentication token for the request. The service supports AAD user/app authentication. For AAD, the header value must be in the format: Bearer {token}.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "The ETag of the previously downloaded Transfer request. A non matching value will result in a failure.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MS-CV",
            "in": "header",
            "description": "Correlation Vector for the request. If not provided, then the service generates a new one.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "The id of the Transfer request to delete.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Empty."
          }
        }
      },
      "get": {
        "tags": [
          "TransferRequest V2"
        ],
        "summary": "Reads a Transfer request by id.",
        "operationId": "getApiV2ById",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authentication token for the request. The service supports AAD user/app authentication. For AAD, the header value must be in the format: Bearer {token}.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MS-CV",
            "in": "header",
            "description": "Correlation Vector for the request. If not provided, then the service generates a new one.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "The id of the Transfer request to retrieve.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "$select",
            "in": "query",
            "description": "A select clause for the query. If provided, then only the requested properties are returned. If not provided, then all properties are returned.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Transfer request if found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.PrivacyServices.DataManagement.DataAccess.Models.V2.TransferRequest"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/TransferRequests": {
      "get": {
        "tags": [
          "TransferRequest V2"
        ],
        "summary": "Reads all Transfer requests using the provided query options.",
        "operationId": "getApiV2TransferRequest",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authentication token for the request. The service supports AAD user/app authentication. For AAD, the header value must be in the format: Bearer {token}.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MS-CV",
            "in": "header",
            "description": "Correlation Vector for the request. If not provided, then the service generates a new one.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "$select",
            "in": "query",
            "description": "A select clause for the query. If provided, then only the requested properties are returned. If not provided, then all properties are returned.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "$filter",
            "in": "query",
            "description": "A filter clause for the query. Filters are supported for the sourceOwnerId and targetOwnerId properties. Ex: $filter=sourceOwnerId eq '{id}'.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "$top",
            "in": "query",
            "description": "Determines page size for the request. If not provided, then server side paging will be used.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "$skip",
            "in": "query",
            "description": "Determines page index for the request. If not provided, then the first page is returned.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A collection of Transfer requests. If server-side paging is triggered, than the nextLink property will be set. Use that to download the next page.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Microsoft.PrivacyServices.DataManagement.DataAccess.Models.V2.TransferRequest"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Microsoft.PrivacyServices.DataManagement.DataAccess.Models.V2.TrackingDetails": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string"
          },
          "updatedOn": {
            "type": "string",
            "format": "date-time"
          },
          "egressedOn": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Microsoft.PrivacyServices.DataManagement.DataAccess.Models.V2.TransferRequest": {
        "type": "object",
        "properties": {
          "sourceOwnerId": {
            "type": "string"
          },
          "targetOwnerId": {
            "type": "string"
          },
          "requestState": {
            "enum": [
              "None",
              "Pending",
              "Approving",
              "Approved",
              "Cancelled",
              "Failed"
            ],
            "type": "string"
          },
          "assetGroups": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id": {
            "type": "string"
          },
          "eTag": {
            "type": "string"
          },
          "trackingDetails": {
            "$ref": "#/components/schemas/Microsoft.PrivacyServices.DataManagement.DataAccess.Models.V2.TrackingDetails"
          }
        }
      },
...
    }
  }
}